+++ ShaderUtils.hpp
@@ -14,6 +14,8 @@
 #ifndef CSCI441_SHADER_UTILS_HPP
 #define CSCI441_SHADER_UTILS_HPP
 
+#include "constants.h"
+
 #ifdef CSCI441_USE_GLEW
     #include <GL/glew.h>
 #else